update: move directory
