+++ components/AgencyBarChart.vue
@@ -49,7 +49,7 @@ import VueI18n from 'vue-i18n'
 import { ChartOptions } from 'chart.js'
 import { ThisTypedComponentOptionsWithRecordProps } from 'vue/types/options'
 import DataView from '@/components/DataView.vue'
-import { getGraphSeriesStyle } from '@/utils/colors'
+import { getGraphSeriesStyle } from '@/libraries/utils/colors'
 import { Registry } from '@/libraries/Registry'
 const agencyData = Registry.AgencyRepository.data
