+++ components/cards/ConfirmedCasesDetailsCard.vue
@@ -28,7 +28,7 @@
 </style>
 
 <script>
-import formatConfirmedCases from '@/utils/formatConfirmedCases'
+import formatConfirmedCases from '@/libraries/utils/formatConfirmedCases'
 import DataView from '@/components/DataView.vue'
 import ConfirmedCasesDetailsTable from '@/components/ConfirmedCasesDetailsTable.vue'
 import { Registry } from '@/libraries/Registry'
