+++ components/cards/ConfirmedCasesNumberCard.vue
@@ -13,7 +13,7 @@
 </template>
 
 <script>
-import formatGraph from '@/utils/formatGraph'
+import formatGraph from '@/libraries/utils/formatGraph'
 import TimeBarChart from '@/components/TimeBarChart.vue'
 import { Registry } from '@/libraries/Registry'
