+++ components/flow/FlowSpElder.vue
@@ -86,7 +86,7 @@
 </template>
 
 <script lang="ts">
-import { onDoneScroll } from '@/utils/vueScrollTo'
+import { onDoneScroll } from '@/libraries/utils/vueScrollTo'
 import AccessibleIcon from '@/static/flow/responsive/accessible.svg'
 import ArrowForwardIcon from '@/static/flow/responsive/arrow_forward.svg'
 import DirectionsWalkIcon from '@/static/flow/responsive/directions_walk.svg'
