+++ components/flow/FlowSpSuspect.vue
@@ -49,7 +49,7 @@
 </template>
 
 <script lang="ts">
-import { onDoneScroll } from '@/utils/vueScrollTo'
+import { onDoneScroll } from '@/libraries/utils/vueScrollTo'
 import ArrowForwardIcon from '@/static/flow/responsive/arrow_forward.svg'
 import PhoneIcon from '@/static/flow/responsive/phone.svg'
 import SentimentIcon from '@/static/flow/responsive/sentiment_very_dissatisfied.svg'
